requirements.txt
addressparser==0.2.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.13
distro==1.9.0
exceptiongroup==1.3.1
faiss-gpu==1.7.2
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jieba==0.42.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
murmurhash==1.0.15
networkx==3.4.2
numpy==1.26.4
openai==2.18.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.3.3
pillow==12.1.1
platformdirs==4.5.1
preshed==3.0.12
prettytable==3.17.0
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
pyecharts==2.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.52.0
simplejson==3.20.2
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
thinc==8.3.10
threadpoolctl==3.6.0
tqdm==4.67.3
typer-slim==0.21.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wandb==0.24.2
wasabi==1.1.3
wcwidth==0.6.0
weasel==0.4.3
wrapt==2.1.1
